Add test for hcx Integrator
